Add slip-n-slide motion player
Bullets are sourced at the player now, too. I don't know why it took so long to
figure out the bullets.

I would like to not have the current force based motion. I'd rather, if
anything, that it was a fast -- but sloped -- acceleration and an instant
deceleration. I'm, also, not at all sure if I should be using a Kinematic or
rigid body for the player.