Add snappy "momentous" motion for player
The motion has a certain 'weight' to its feeling. (No realistic physics is
happening, though. Please bear that in mind.) The acceleration is quick,
though, which is good for a bullet hell game with perhaps a measure of
dodging involved. The weight to your motion, though, provides what feels like a
bit of a backpedal almost. Letting yourself stop is gentle and slow, which
feels, at least somewhat, realistic.