Add Dumb Enemy "AI"
It just walks straight at the player. No path planning whatsoever.

I'll only add path planning depending on what the maps/enemies end up being.
If there are bosses, I'll do this sort of boid behaviour. If there are nooks
and crannies on the maps, I'll add some smarter global planning for the
individuals. What I might do is a sort of "secret" boss: a path that all of the
enemies follow.