Add scenes for map scrolling
No functionality as of now

I'll be implementing "idea 1" where the map has a backdrop and mostly scrolls
left and right. (Theoretically, up and down is possible, but that would be an
actual scene transition.)

The main play area will exist in front of the backdrop, and it will be
represented as a tilemap of ground sprites.

There will be objects that go on the main arena/courtyard/road/yard such as
stalls, carts, fountains, and etc.